Adventure 2.0
Program utilizing functions for an "adventure" where the player is asked if it wishes to go on an adventure or not, and then generates attack and defense numbers, which determines the outcome of a series of fights, and awards the user treasure, and damages the player, accordingly. This continues until either the player says so, or the player dies, in which the program ends and alerts the player of their results.